+++ medcom-video-api-service/src/main/java/dk/medcom/video/api/dao/EntitiesIvrThemeDaoImpl.java
@@ -1,4 +1,4 @@
-package dk.medcom.video.api.dao.impl;
+package dk.medcom.video.api.dao;
 
 import dk.medcom.video.api.dao.EntitiesIvrThemeDao;
 import dk.medcom.video.api.dao.entity.EntitiesIvrTheme;
