+++ medcom-video-api-service/src/main/java/dk/medcom/video/api/service/MeetingUserServiceImpl.java
@@ -1,11 +1,9 @@
-package dk.medcom.video.api.service.impl;
+package dk.medcom.video.api.service;
 
 import dk.medcom.video.api.context.UserContextService;
 import dk.medcom.video.api.controller.exceptions.PermissionDeniedException;
 import dk.medcom.video.api.dao.MeetingUserRepository;
 import dk.medcom.video.api.dao.entity.MeetingUser;
-import dk.medcom.video.api.service.MeetingUserService;
-import dk.medcom.video.api.service.OrganisationService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
