+++ medcom-video-api-service/src/main/java/dk/medcom/video/api/service/MessagingException.java
@@ -1,4 +1,4 @@
-package dk.medcom.video.api.service.impl;
+package dk.medcom.video.api.service;
 
 public class MessagingException extends RuntimeException {
     public MessagingException(String message, Exception e) {
