+++ medcom-video-api-service/src/main/java/dk/medcom/video/api/service/NewProvisionerOrganisationFilterImpl.java
@@ -1,4 +1,4 @@
-package dk.medcom.video.api.service.impl;
+package dk.medcom.video.api.service;
 
 import dk.medcom.video.api.service.NewProvisionerOrganisationFilter;
