+++ medcom-video-api-service/src/main/java/dk/medcom/video/api/service/OrganisationServiceImpl.java
@@ -1,11 +1,10 @@
-package dk.medcom.video.api.service.impl;
+package dk.medcom.video.api.service;
 
 import dk.medcom.video.api.context.UserContextService;
 import dk.medcom.video.api.controller.exceptions.PermissionDeniedException;
 import dk.medcom.video.api.dao.OrganisationRepository;
 import dk.medcom.video.api.dao.entity.Organisation;
 import dk.medcom.video.api.organisation.OrganisationStrategy;
-import dk.medcom.video.api.service.OrganisationService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
