+++ medcom-video-api-service/src/main/java/dk/medcom/video/api/service/PoolFinderServiceImpl.java
@@ -1,11 +1,10 @@
-package dk.medcom.video.api.service.impl;
+package dk.medcom.video.api.service;
 
 import dk.medcom.video.api.api.CreateMeetingDto;
 import dk.medcom.video.api.api.ProvisionStatus;
 import dk.medcom.video.api.dao.SchedulingInfoRepository;
 import dk.medcom.video.api.dao.entity.Organisation;
 import dk.medcom.video.api.dao.entity.SchedulingInfo;
-import dk.medcom.video.api.service.PoolFinderService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
