+++ medcom-video-api-service/src/test/java/dk/medcom/video/api/service/impl/AuditServiceImplTest.java
@@ -4,6 +4,7 @@ import dk.kvalitetsit.audit.client.AuditClient;
 import dk.kvalitetsit.audit.client.api.v1.AuditEvent;
 import dk.medcom.video.api.api.*;
 import dk.medcom.video.api.dao.entity.*;
+import dk.medcom.video.api.service.AuditServiceImpl;
 import org.junit.Before;
 import org.junit.Test;
 import org.mockito.ArgumentCaptor;
