+++ medcom-video-api-service/src/test/java/dk/medcom/video/api/service/impl/MeetingServiceImplTest.java
@@ -15,10 +15,7 @@ import dk.medcom.video.api.dao.OrganisationRepository;
 import dk.medcom.video.api.dao.entity.*;
 import dk.medcom.video.api.organisation.model.OrganisationTree;
 import dk.medcom.video.api.organisation.OrganisationTreeServiceClient;
-import dk.medcom.video.api.service.AuditService;
-import dk.medcom.video.api.service.MeetingService;
-import dk.medcom.video.api.service.OrganisationService;
-import dk.medcom.video.api.service.SchedulingInfoEventPublisher;
+import dk.medcom.video.api.service.*;
 import org.hibernate.exception.ConstraintViolationException;
 import org.junit.Assert;
 import org.junit.Before;
