+++ medcom-video-api-service/src/test/java/dk/medcom/video/api/service/impl/OrganisationFinderTest.java
@@ -1,6 +1,7 @@
 package dk.medcom.video.api.service.impl;
 
 import dk.medcom.video.api.organisation.model.OrganisationTree;
+import dk.medcom.video.api.service.OrganisationFinder;
 import org.junit.Before;
 import org.junit.Test;
