+++ medcom-video-api-service/src/test/java/dk/medcom/video/api/service/impl/PoolFinderServiceTest.java
@@ -6,6 +6,7 @@ import dk.medcom.video.api.api.VmrQuality;
 import dk.medcom.video.api.dao.SchedulingInfoRepository;
 import dk.medcom.video.api.dao.entity.Organisation;
 import dk.medcom.video.api.dao.entity.SchedulingInfo;
+import dk.medcom.video.api.service.PoolFinderServiceImpl;
 import org.junit.Before;
 import org.junit.Test;
 import org.mockito.Mockito;
