+++ medcom-video-api-service/src/test/java/dk/medcom/video/api/service/impl/SchedulingTemplateServiceTest.java
@@ -9,7 +9,9 @@ import dk.medcom.video.api.api.DirectMedia;
 import dk.medcom.video.api.controller.exceptions.NotAcceptableException;
 import dk.medcom.video.api.organisation.model.OrganisationTree;
 import dk.medcom.video.api.organisation.OrganisationTreeServiceClient;
+import dk.medcom.video.api.service.MeetingUserServiceImpl;
 import dk.medcom.video.api.service.OrganisationService;
+import dk.medcom.video.api.service.SchedulingTemplateServiceImpl;
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
